Add files via upload
Implemented a better fix for time bases when extracting cuesheets. Sometimes the timebase of the embedded chapters does not match the timebase of the file. This change should account for that. Note that these scripts choose to write chapters with a default 1/1000 time base, which may differ from any previously embedded chapter's timebase (for example Libation embeds chapters with 1/44100 or 1/22050 time base).